refactor: rename optima_sync to volunteer_management_system_sync
Rename Lambda handler directory and update all internal references
from "optima" to "volunteer_management_system" to use a generic name
for the external volunteer management system integration.
